add skillsRoutes
